requirements.txt
Appium-Python-Client==0.50
decorator==4.4.2
docutils==0.16
kitchen==1.2.6
Pygments==2.6.1
robotframework==3.1.2
robotframework-appiumlibrary==1.5.0.4
robotframework-datadriver==0.3.6
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.3.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
